pyproject.toml
[project]
name = "floatchat"
version = "0.1.0"
description = "An AI assistant that turns oceanographic data into answers. Skip complex portals and formats—just ask, “Show warm water near Madagascar in 2023” or “Compare salinity in the Arabian Sea vs Bay of Bengal”—and get instant results with clear visualizations."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-experimental>=0.3.4",
    "langchain-groq>=0.3.8",
    "netcdf4>=1.7.2",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "pathlib>=1.0.1",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "rich>=14.1.0",
    "sqlalchemy>=2.0.43",
    "tqdm>=4.67.1",
]
